requirements.txt
configparser
flickrapi==2.3.1
setuptools
python-magic

